transfer: rename "stream" to "abc" because it contains abstract classes for both Stream and Connection
